requirements.txt
fastapi==0.92.0
python-dotenv==0.21.0
datadog-api-client==2.10.0